% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpbart.R
\name{gp_bart}
\alias{gp_bart}
\title{Fit a GP-BART model}
\usage{
gp_bart(
  x,
  y,
  number_trees = 2,
  node_min_size = 15,
  mu = 0,
  alpha = 0.5,
  beta = 5,
  tau = 1,
  n_iter = 2000,
  burn = 500,
  thin = 1,
  rotation = TRUE,
  theta = NULL,
  seed = NULL,
  scale_boolean = TRUE,
  nu_vector = NULL,
  a_tau = 3,
  d_tau = 1,
  discrete_phi_boolean = FALSE,
  x_scale = TRUE,
  gp_variables = colnames(x),
  K_bart = 2,
  prob_tau = 0.9,
  kappa = 0.5,
  bart_boolean = TRUE,
  bart_number_iter = 250
)
}
\arguments{
\item{x}{A named matrix x of the covariates from the training data.}

\item{y}{A named matriy y of the response from the training data}

\item{number_trees}{Number of trees used in the GP-BART model.}

\item{node_min_size}{Minimum number of observations in a terminal node}

\item{mu}{Initial values for the \eqn{\mu} parameter}

\item{alpha}{Parameter \eqn{\alpha} from tree prior}

\item{beta}{Parameter \eqn{\beta} from tree prior}

\item{tau}{Initial value for \eqn{\tau} parameter}

\item{n_iter}{Total number of iterations in GP-BART}

\item{burn}{Number of burn MCMC iterations}

\item{thin}{Number of thinning from MCMC}

\item{rotation}{Boolean to decide if rotated splits will be used}

\item{theta}{Fixed value for \eqn{\theta}, if NULL a random value for \eqn{\theta} will be used.}

\item{seed}{Setting a seed for the model}

\item{scale_boolean}{Boolean to scale or not the response \eqn{y}}

\item{a_tau}{Scale parameter from \eqn{\tau} prior}

\item{d_tau}{Rate parameter from \eqn{\tau} prior}

\item{discrete_phi_boolean}{Boolean to decide if it will be used a discrete grid for \eqn{\phi} proposals}

\item{x_scale}{Boolean to scale x or not}

\item{gp_variables}{Covariates used to build the covariance matrix from GP}

\item{K_bart}{Prior parameter from \eqn{\tau_{\mu}}}

\item{prob_tau}{Prior parameter from \eqn{\tau}}

\item{kappa}{Weight parameter \eqn{\kappa} from the proposed mixture of BART and GP-BART}

\item{bart_boolean}{Boolean to decide or not to warmup the model with BART samples}

\item{bart_number_iter}{Number of BART initial iterations to be used}
}
\value{
A 'gpbart_GPBART' model object,
}
\description{
Fit a GP-BART model
}
