\name{GPBart-package}
\alias{GPBart-package}
\alias{GPBart}
\docType{package}
\title{
\packageTitle{GPBart}
}
\description{
\packageDescription{GPBart}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GPBart}
\packageIndices{GPBart}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GPBart}

Maintainer: \packageMaintainer{GPBart}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
