% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bart.R
\name{gpbart}
\alias{gpbart}
\title{GP-BART: Gaussian Processes Bayesian Additive Regression Trees}
\usage{
gpbart(
  x_train,
  y,
  x_test,
  n_tree = 20,
  node_min_size = 2,
  n_mcmc = 3500,
  n_burn = 1500,
  alpha = 0.95,
  beta = 2,
  df = 3,
  sigquant = 0.9,
  kappa = 2,
  tau = 100,
  scale_bool = TRUE,
  nu = 1,
  rand_tau_init = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x_train}{Set of explanatory variables of the training data. It must be a \code{data.frame()}}

\item{y}{Response variable for the training data}

\item{x_test}{Set of explanatory variables of the test data. It must be a \code{data.frame()}}

\item{n_tree}{Number of Trees used in the GP-BART model.}

\item{node_min_size}{Node minimum size of observations within a terminal node}

\item{n_mcmc}{The number of MCMC iterations}

\item{n_burn}{The number of MCMC iterations to be trated as burn in}

\item{alpha}{Base parameter for the tree prior}

\item{beta}{Power parameter for the tree prior}

\item{df}{Degrees of freedom for the residual precision prior}

\item{sigquant}{The quantile of the residual precision prior;}

\item{kappa}{The number of prior standard deviations away from the from the range of the response.}

\item{tau}{Initial value for the residual precision}

\item{scale_bool}{A Boolean to choose if will be scaled or not.}

\item{nu}{Value for the GP precision. The default value is \eqn{\nu = 4\kappa^{2}T}}

\item{rand_tau_init}{A Boolean to let the initial value be initialised or not. The default is \code{TRUE}.}

\item{verbose}{Verbosity flag for printing progress. The default is \code{TRUE}}
}
\description{
GP-BART is an extension to to the Bayesian Additive Regression Trees (BART)
}
