% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{k_fold}
\alias{k_fold}
\title{K-fold}
\usage{
k_fold(
  data,
  dependent_variable = "y",
  k_partitions,
  seed = NULL,
  as_data_frame = FALSE
)
}
\arguments{
\item{data}{A data.frame with the whole data to be split into training and test folds.}

\item{dependent_variable}{A string with the name of the target variable}

\item{k_partitions}{Number of folds to be used in the K-Fold setting}

\item{seed}{Setting a seed to reproducibility of the cross validation}

\item{as_data_frame}{A Boolean to return the \code{x_train} and \code{x_test} partitions as \code{data.frame}}
}
\description{
A K-fold setting for crossvalidation
}
