% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{A numerical vector of observed variables}

\item{y}{A numerical vector of predicted variables}
}
\description{
Calculating the Root Mean Squared Error
}
